PublicSuffix::List#select is now private
Note that this method is currently private, as you should not rely on
it. Instead, the public interface is {#find}. The current internal
algorithm allows to return all matching rules, but different data
structures may not be able to do it, and instead would return only the
match.

For this reason, you should rely on {#find}.