Change lookup algorithm to Hash
The Hash doesn't require manual reindexing when new rules are added.
Moreover, the Hash-based algorithm has almost O(1) lookup time.

Actually, the lookup time is O(k), where k is the number of parts in
the input string.

    find("www.example.com") -> k = 2
    find("www.example.com") -> k = 3
    find("www.subdomain.example.com") -> k = 4

It's fair to consider that the average number of parts is 3, and
hostnames longer than 5 parts are quite uncommon.

Note that the Hash-based lookup is highly influenced by whatever
underlying Hash implementation is provided by the programming language.
A Perfect Hash would be preferable in terms of lookup time as it offers
real O(1) lookup time complexity (whereas a dynamic Hash is avg O(1)),
however a Perfect Hash would require a computation of a perfect hashing
function, without considering that it would not allow the flexibility
of adding/removing rules at runtime.