PublicSuffix::List is no longer enumerable
Although the list is still a collection of rules, the idea is that the
internal list representation is shifting away from a simple collection,
and looping over rules is not only discouraged but it may actually be
impossible in the future.

Moreover, including Enumerable will introduce a bunch of methods.

Instead, if you really need to iterate over the collection, simply use
#each that returns an Enumerator.